refactor: renames districts file
