+++ src/scraping/main.py
@@ -1,3 +1,3 @@
-from listing_districts import districts
+from districts import districts
 
 print(districts)
